Remove GradedAxesNext
