+++ NDTensors/src/imports.jl
@@ -37,7 +37,6 @@ for lib in [
   :RankFactorization,
   :Sectors,
   :LabelledNumbers,
-  :GradedAxesNext,
   :GradedAxes,
   :TensorAlgebra,
   :SparseArrayInterface,
