--- NDTensors/src/lib/GradedAxes/ext/GradedAxesSectorsExt/Project.toml
@@ -1,2 +1,0 @@
-[deps]
-BlockArrays = "8e7c35d0-a365-5155-bbbb-fb81a777f24e"
