+++ NDTensors/src/lib/GradedAxes/src/GradedAxes.jl
@@ -1,7 +1,4 @@
 module GradedAxes
-include("groupsortperm.jl")
-include("tensor_product.jl")
-include("abstractgradedunitrange.jl")
 include("gradedunitrange.jl")
-include("../ext/GradedAxesSectorsExt/src/GradedAxesSectorsExt.jl")
+include("fusion.jl")
 end
