--- NDTensors/src/lib/GradedAxesNext/.JuliaFormatter.toml
@@ -1,2 +1,0 @@
-style = "blue"
-indent = 2
