--- NDTensors/src/lib/GradedAxesNext/src/GradedAxesNext.jl
@@ -1,3 +1,0 @@
-module GradedAxesNext
-include("gradedunitrange.jl")
-end
