--- NDTensors/src/lib/GradedAxesNext/test/Project.toml
@@ -1,4 +1,0 @@
-[deps]
-BlockArrays = "8e7c35d0-a365-5155-bbbb-fb81a777f24e"
-NDTensors = "23ae76d9-e61a-49c4-8f12-3f1a16adf9cf"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
