+++ NDTensors/src/lib/TensorAlgebra/ext/TensorAlgebraGradedAxesExt/src/TensorAlgebraGradedAxesExt.jl
@@ -1,8 +1,8 @@
 module TensorAlgebraGradedAxesExt
-using ...GradedAxes: AbstractGradedUnitRange, tensor_product
+using ...GradedAxes: GradedUnitRange, tensor_product
 using ..TensorAlgebra: TensorAlgebra
 
-function TensorAlgebra.:⊗(a1::AbstractGradedUnitRange, a2::AbstractGradedUnitRange)
+function TensorAlgebra.:⊗(a1::GradedUnitRange, a2::GradedUnitRange)
   return tensor_product(a1, a2)
 end
 end
