+++ NDTensors/test/lib/runtests.jl
@@ -10,7 +10,6 @@ using Test: @testset
   "CUDAExtensions",
   "DiagonalArrays",
   "GradedAxes",
-  "GradedAxesNext",
   "GPUArraysCoreExtensions",
   "LabelledNumbers",
   "MetalExtensions",
